Further refactoring of test classes and changed Junit4 to Junit5
