+++ src/test/java/org/broadleafcommerce/frameworkmapping/DisabledFrameworkMappingTest.java
@@ -3,22 +3,23 @@ package org.broadleafcommerce.frameworkmapping;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
 import org.broadleafcommerce.frameworkmapping.annotation.FrameworkController;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.FilterType;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest(includeFilters = {@ComponentScan.Filter(classes = FrameworkController.class),
                               @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE,
                                                     classes = FrameworkControllerHandlerMapping.class)})
 @ContextConfiguration(classes = DisabledFrameworkApplication.class)
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 public class DisabledFrameworkMappingTest {
 
     @Autowired
