+++ src/test/java/org/broadleafcommerce/frameworkmapping/support/CustomController.java
@@ -1,4 +1,4 @@
-package org.broadleafcommerce.frameworkmapping;
+package org.broadleafcommerce.frameworkmapping.support;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
